feat: create common entry file
