+++ packages/author-profile/author-profile.js
@@ -1,10 +1,7 @@
 import React from "react";
-import { Text, FlatList, Linking } from "react-native";
+import { Text } from "react-native";
 import PropTypes from "prop-types";
-import AuthorProfileFooter from "./author-profile-footer";
-import AuthorProfileHeader from "./author-profile-header";
-import AuthorProfileItem from "./author-profile-item";
-import AuthorProfileItemSeparator from "./author-profile-item-separator";
+import AuthorProfileData from "./author-profile-data";
 
 const AuthorProfile = props => {
   const data = props.data;
@@ -12,36 +9,16 @@ const AuthorProfile = props => {
     return <Text>Loading</Text>;
   }
 
-  const headerProps = {
-    ...data.author,
-    count: data.author.articles.count,
-    onNext: url => Linking.openURL(url),
-    onPrev: url => Linking.openURL(url),
-    pageSize: props.pageSize,
-    page: props.page
-  };
-
-  return (
-    <FlatList
-      data={data.author.articles.list}
-      ItemSeparatorComponent={() => <AuthorProfileItemSeparator />}
-      keyExtractor={article => article.id}
-      ListFooterComponent={() => <AuthorProfileFooter />}
-      ListHeaderComponent={() => <AuthorProfileHeader {...headerProps} />}
-      renderItem={({ item }) => <AuthorProfileItem {...item} />}
-    />
-  );
+  return <AuthorProfileData {...props} />;
 };
 
 AuthorProfile.propTypes = {
-  page: PropTypes.number,
-  pageSize: PropTypes.number,
-  data: PropTypes.shape()
+  data: PropTypes.shape({
+    loading: PropTypes.boolean
+  })
 };
 
 AuthorProfile.defaultProps = {
-  page: 1,
-  pageSize: 10,
   data: {
     loading: true
   }
